Remove nanosleep: It's unreliable in conjunction with faketime. Manually force idle if you actually want to sleep.
We presumed that when a process entered nanosleep, they intended to wait for an internal event: that is, faketime being advanced. However, libfaketime just returns from nanosleep immediately. This hid an issue with the D runtime, where a GC spinlock used nanosleep immediately, *not* intending to put the thread to sleep in a libidle sense.
Since we cannot separate domain sleeps from runtime sleeps, you are asked to manually force the thread to idle if you are waiting for an clock change rather than badly implementing busyspin.